Offline beer browsing
Enable beer browsing when offline.

The app will ship with a static snapshot of the beers in the database.
The static snapshot will be loaded into local storage.
Local storage will be updated every time the list of beers is refreshed
from the server.

When offline, the beer browser will display from the copy in local
storage if refreshed.