Update mybeerlist without full refresh
There was a complaint about the beerlist jumping back to top after
using the Drink button.  It was caused when I added back the refresh()
to get the progress bar to update.  I took that back and now the code
manually increments the number ordered on the beer list and
recalculates the percentage  on the fly.